# api/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "helloworld/api/api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/adduser": {
            "get": {
                "summary": "增删改查",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.Response"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "nickname",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/deleteuser": {
            "get": {
                "summary": "/deleteuser",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.Response"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/getuser": {
            "get": {
                "summary": "/getuser",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.Response"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/getuserlist": {
            "get": {
                "summary": "/getuserlist",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.Response"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/kratos-demo/say_hello": {
            "get": {
                "summary": "/kratos-demo/say_hello",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.HelloResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/updateuser": {
            "get": {
                "summary": "/updateuser",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.Response"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/user/login": {
            "get": {
                "summary": "----------------------------自定义接口-------------------//",
                "description": "新增登录服务接口",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.LoginResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "passwd",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        }
    },
    "definitions": {
        ".demo.service.v1.AddReq": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                }
            },
            "description": "增删改查接口响应参数",
            "required": [
                "nickname",
                "age"
            ]
        },
        ".demo.service.v1.DeleteReq": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "integer"
                }
            },
            "required": [
                "uid"
            ]
        },
        ".demo.service.v1.GetReq": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "integer"
                }
            },
            "required": [
                "uid"
            ]
        },
        ".demo.service.v1.HelloReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            },
            "description": "---------------------------------下面这个不是-----------------------------------------//\n定义请求参数",
            "required": [
                "name"
            ]
        },
        ".demo.service.v1.HelloResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            },
            "description": "定义响应参数"
        },
        ".demo.service.v1.LoginReq": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "passwd": {
                    "type": "string"
                }
            },
            "description": "--------------------自定义message--------------------------//",
            "required": [
                "username",
                "passwd"
            ]
        },
        ".demo.service.v1.LoginResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            },
            "description": "新增登录接口响应参数"
        },
        ".demo.service.v1.Response": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.UpdateReq": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                }
            },
            "required": [
                "uid",
                "nickname",
                "age"
            ]
        },
        ".google.protobuf.Empty": {
            "type": "object",
            "properties": {},
            "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }\n\n The JSON representation for `Empty` is empty JSON object `{}`."
        }
    }
}